phildailybot/requirements.txt
aiohttp==3.5.4
asn1crypto==0.24.0
async-timeout==3.0.1
attrs==18.2.0
beautifulsoup4==4.7.1
certifi==2018.11.29
cffi==1.12.0
chardet==3.0.4
cryptography==2.5
future==0.17.1
idna==2.8
lxml==4.3.1
multidict==4.5.2
pycparser==2.19
python-telegram-bot==11.1.0
requests==2.21.0
six==1.12.0
soupsieve==1.7.3
telepot==12.7
urllib3==1.24.1
yarl==1.3.0
